KEIJO-157: Combine FormComboBox into JiraIssueComboBox
